KS-2-2 the solution to lesson 2 task 2, corrected

+++ src/main/kotlin/Lesson1/lesson1-task4.kt
@@ -1,10 +1,10 @@
 package org.example.Lesson1
 
-fun main (){
+fun main() {
     val distance: Long = 4086860000
     val age: Byte = 27
     val partOfDay: Float = 0.075f
-    val seconds: Int = 6480
+    val seconds: Short = 6480
     val partOfTheYear: Double = 2.0547945205479453E-4
     val apogeeOfTheOrbit: Int = 327000
